feat: start a bit of refactoring
Multiobjectives still dont work consistently...
